Fix the issue that the maximum width of the column does not change when a format is applied to the column
When the getmaxwidth() function calculates the maximum width of a column, the format is not applied to the column. When the user sets the format, the maximum number of columns that the Dataset can display is still calculated according to the original data. However, the maximum number of displayed columns may change after setting the format.

Now the format of each column will be applied before calculating the maximum width of the column. The maximum number of columns that can be displayed will be changed correctly.