Fix the issue that the maximum width of the column does not change when a format is applied to the column (#86)
* Improve the speed of displaying large datasets in Jupyter Notebook

Fix the issue that Jupyter Notebook takes too much time when displaying large datasets with many columns.

The getmaxwidths() function now only calculates the maximum width of the columns that will be displayed, not all columns.

Please see issue for more details.

* Modify the initialization code of maxwidths

Now might break the loop before filling all the values of maxwidths.

The initialize code of maxwidths has been modified.

* Fix the issue that the maximum width of the column does not change when a format is applied to the column

When the getmaxwidth() function calculates the maximum width of a column, the format is not applied to the column. When the user sets the format, the maximum number of columns that the Dataset can display is still calculated according to the original data. However, the maximum number of displayed columns may change after setting the format.

Now the format of each column will be applied before calculating the maximum width of the column. The maximum number of columns that can be displayed will be changed correctly.

* Add mapformats keyword argument to getmaxwidths() and _show()

The FOR loop is duplicated so that mapformats only need to be determined once.

* Update io.jl

Simplify part of the code of getmaxwidths()

* Update io.jl

Make mapformats as a keyword argument and update the docstring of the getmaxwidths() function.